requirements.txt
aioredis==1.3.1
argon2-cffi==20.1.0
asgi-redis==1.4.3
asgiref==3.2.10
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.5
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
compose==1.1.1
configparser==5.0.0
constantly==15.1.0
crayons==0.3.1
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
dj-database-url==0.5.0
Django==2.0.2
django-catalog==2.1.1
django-ckeditor==5.9.0
django-classy-tags==2.0.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-recaptcha==2.0.6
django-storages==1.10.1
djangorestframework==3.11.1
docker==4.2.2
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
extdirect.django==0.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
future==0.17.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==19.0.0
idna==2.6
importlib-metadata==1.7.0
incremental==17.5.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
MarkupSafe==1.1.1
mistune==0.8.4
modules==1.0.0
msgpack==1.0.0
msgpack-python==0.5.6
mysql-connector==2.2.9
nbclient==0.5.0
nbconvert==6.0.2
nbformat==5.0.7
nest-asyncio==1.4.0
newsapi-python==0.2.6
notebook==6.1.4
numpy==1.20.1
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.1
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.7.0
PyHamcrest==2.0.2
PyMySQL==0.9.3
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pyttsx3==2.71
pytz==2019.3
pywin32==224
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
redis==2.10.6
requests==2.24.0
scipy==1.6.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
tornado==6.0.4
traitlets==5.0.4
Twisted==20.3.0
txaio==20.4.1
urllib3==1.22
ven==0.4.1
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
wcwidth==0.2.5
webdriver-manager==3.2.1
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
zipp==3.1.0
zope.interface==5.1.0
